requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.2
gunicorn==20.1.0
importlib-metadata==4.0.1
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
whitenoise==5.2.0
zipp==3.4.1
